Move activity / remove unused code / add support for back navigation all the way down to api level 16
